Fixes a bug introduced in ThreadIsolatedValue by my previous commit
Updates the NAnt build script so that it will now sign most of our assemblies.  Silverlight and Windows Phone 7 builds are not signed because those platforms have known problems with interactions between strong named assemblies and NuGet.  Also the Query.FullText library is not currently signed for .Net 3.5 because the old Lucene.Net compatible with .Net 3.5 is not itself signed
Affects Issue [273]