Normalize datatypes before comparing graphs
In the RDF 1.1 model, literals with no language tag and no datatype are
treated as having a default datatype of xsd:string. The JSON-LD parser
enforces this, but other DNR parsers do not. As a result we get some
graph comparison errors in these tests. This fix "normalizes" all graphs
so that literal nodes with no language tag or datatype URI get an
xsd:string datatype.