Fix to IRI formatting for NTriples and NQuads output
This fix replaces the %-encoding that was taking place with \uXXXX
encoding of only the reserved characters specified in the
NTriples/NQuads spec. Due to the way the .NET Uri class constructor
works, this code still fails to properly format URIs that contain the %
character but this could only be fixed by eliminating use of the .NET
Uri class from the codebase.