Add SingleGraphWriter helper class
Adds SingleGraphWriter as a convenience class that wraps an IStoreWriter
with an IRdfWriter interface, making it easier to write a single graph
to a writer that implements the IStoreWriter interface.

To fully implement the IRdfWriter interface it was necessary to add an
override to the IStoreWriter interface that allows the caller to specify
that the output stream should be left open after the store is saved.