Catch weird CI exception
On the CI server (and not on my local machine) the Uri.ToString method
throws an ArgumentOutOfRangeException for a small number of the unit
tests. This is just in code that is writing debug output to the console
so for now I've wrapped the code affected in a try/catch to suppress the
exception.