Improve XML namespace handling in RDF/XML parser
Reworked the RDFXMLParser and RdfXmlSpecsHelper code to better handle the mapping of the RDF namespace to prefixes other than "rdf".
The rework is a bit hampered by wanting to preserve existing APIs. The pre-existing APIs in RdfXmlSpecsHelper will not properly handle namespace mappings are are now marked as deprecated for this reason with new implementations provided as an upgrade path. The internal usage of these APIs has been updated.