Make RDF/JSON parser more lenient
Now allows the value property to be a number as well as a string. Raises a warning if the value property is not a string (as this is not strictly valid by the spec), but will convert the value to a string using the InvariantCulture formatter.
This allows the parser to handle the (broken) RDF/JSON that the DBPedia endpoint is currently producing.