Feature/issue 258 (#259)
* Support overriding the format to use when sending RDF to the server

Fixes #258

* Rewrite skip to use xunit.skippablefact package

* Update TestConfigManager for new skippable fact infrastructure

* Do not attempt to run async begininvoke methods under .NET Core

* Make RDF/JSON parser more lenient

Now allows the value property to be a number as well as a string. Raises a warning if the value property is not a string (as this is not strictly valid by the spec), but will convert the value to a string using the InvariantCulture formatter.
This allows the parser to handle the (broken) RDF/JSON that the DBPedia endpoint is currently producing.