Fix GraphML output and test
The GraphML output was not valid for named graphs as it was trying to put the graph URI into the id attribute of the graph element (which the schema has defined as an XML ID type attribute).
This patch adds a new graphlabel data attribute which is used to hold the graph URI for named graphs in the GraphML output.