Remove use of static Options.UseBomForUtf8 and Options.ForceHttpBasicAuth
For #310

All writers now default to encoding files in UTF-8 with no BOM and provide an overload to the Save() method that takes a specific text encoding to use.

All web requests now default to the .NET Framework handling for credentials. This will be revisited when those APIs are updated to use HttpClient as part of #309.