Completed removal of static Options class
Moved `Options.LiteralEqualityMode` to the `VDS.RDF.EqualityHelper` static class. This option remains a static option as there isn't a simple way to inject it into the .Equals overrides.

Moved `Options.NormalizeLiteralValues` to the `VDS.RDF.INodeFactory` interface so that this option can now be set on the factory (or graph) used to create nodes.
#310