Ensure that GROUP BY does not create an implicit group when the input… (#352)
* Update groupby implementation and tests

The tests are now aligned with the recent updates to the SPARQL 1.1 query test suite with respect to the handling of aggregates with empty groups. If the input is empty and an explicit grouping is specified, the output is empty. If the input is empty and there are aggregates but no explicit grouping the output contains a single row.

Fixes #351