Update properties that expose graph names
Updated properties in the storage, store interfaces and query interfaces so that graphs can be referenced using IRefNodes rather than URIs. URI-based interfaces are all deprecated and marked as obsolete.
Lots of internal fixes so that all unit tests pass.
At this stage there are still a lot of internal uses of deprecated methods to be tidied up.