Refactored Describe algorithm support
Modified the Describe algorithm interface to be independent of the SPARQL evaluation context. The describe algorithms are now generic enough to be applied to a list of nodes and a given graph/dataset.
Extracted a common interface for triple lookup (ITripleIndex) so that describe algorithms do not have to be separately implemented for ISparqlDataset and IGraph.
This is a breaking change. The API for the algorithms has been modified and the algorithms have now moved to VDS.RDF.Utils.Describe. The wrapper class VDS.RDF.Query.SparqlDescriber is used to adapt an IDescribeAlgorithm to work with a Leviathan SparqlEvaluationContext.