Support overriding the default XML Reader settings for XML parsers
The TriX and RDF/XML parsers now expose an XmlReaderSettings property which can be used to provide the settings that will be passed to the underlying XML parser when a Load() method is invoked. The default settings remain the same as in 2.x, but it is now possible to disable DTD and PI processing.